requirements.txt
streamlit
openai
PyMuPDF
python-dotenv
pinecone
tiktoken
